helm: use port 80/443 by default for the peer service
When the service port for the peer service is not specified, it is
automatically assigned port 80 or port 443 (when TLS is enabled). Using
these ports make it easy to understand whether TLS is enabled for the
service or not. Moreover, it makes the behavior consistent with the
Hubble Relay service.

Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>